% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{blogdown}
\alias{blogdown}
\alias{blogdown-package}
\title{The \pkg{blogdown} package}
\description{
The comprehensive documentation of this package is the book \bold{blogdown:
Creating Websites with R Markdown}
(\url{https://bookdown.org/yihui/blogdown/}). You are expected to read at
least the first chapter. If you are really busy or do not care about an
introduction to \pkg{blogdown} (e.g., you are very familiar with creating
websites), set your working directory to an empty directory, and run
\code{blogdown::\link{new_site}()} to get started right away.
}
\examples{
if (interactive()) blogdown::new_site()
}
