% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{build_site}
\alias{build_site}
\title{Build a website}
\usage{
build_site(local = FALSE)
}
\arguments{
\item{local}{Whether to build the website locally to be served via
\code{\link{serve_site}()}. If \code{TRUE}, the site configurations
\code{baseurl} will be set to \code{/}, and \code{relativeurls} will be set
to \code{true}. If \code{FALSE}, default configurations of the website will
be used.}
}
\description{
Compile all Rmd files, build the site through Hugo, and post-process HTML
files generated from Rmd (e.g. fix figure paths).
}
\details{
You can use \code{\link{serve_site}()} to preview your website locally, and
\code{build_site()} to build the site for publishing.
}
\note{
When \code{local = TRUE}, RSS feeds (typically the files named
  \code{index.xml} under the \file{public} directory) will not be
  post-processed to save time, which means if you have Rmd posts that contain
  R plots, these plots will not work. Since \code{local = TRUE} is only for
  previewing a website locally, you may not care about RSS feeds.
}
