% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site_modified.R
\name{build_site_modified}
\alias{build_site_modified}
\title{Build a site, using only modified .Rmd files}
\usage{
build_site_modified()
}
\description{
The function \code{build_site_modified()} builds the site through Hugo,
using only modified .Rmd files. This avoids the case where blogdown
compiles all .Rmd files in the website directory, which can often cause
problems with code from an earlier date that does not run anymore for whatever
reason.
}
