% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{clean_duplicates}
\alias{clean_duplicates}
\title{Clean duplicated output files}
\usage{
clean_duplicates(preview = TRUE)
}
\arguments{
\item{preview}{Whether to preview the file list, or just delete the files. If
you are sure the files can be safely deleted, use \code{preview = FALSE}.}
}
\value{
For \code{preview = TRUE}, a logical vector indicating if each file
  was successfully deleted; for \code{preview = FALSE}, the file list is
  printed.
}
\description{
For an output file \file{FOO.html}, \file{FOO.md} should be deleted if
\file{FOO.Rmd} exists, and \file{FOO.html} should be deleted when
\file{FOO.Rmarkdown} exists (because \file{FOO.Rmarkdown} should generate
\file{FOO.markdown} instead) or neither \file{FOO.Rmarkdown} nor
\file{FOO.Rmd} exists (because a plain Markdown file should not be knitted to
HTML).
}
