% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_helper.R
\name{config_helper}
\alias{config_helper}
\title{Config Finder for Complex Themes
Some themes, like the popular wowchemy academic themes, place the config.toml or
config.yaml to unusual paths. This helper copies the config to the project root.}
\usage{
config_helper(config_dir = file.path("config", "_default"))
}
\arguments{
\item{config_dir}{Defaults to the wowchemy template defaults \code{ file.path( "config", "_default")}}
}
\description{
Config Finder for Complex Themes
Some themes, like the popular wowchemy academic themes, place the config.toml or
config.yaml to unusual paths. This helper copies the config to the project root.
}
