% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{config_Rprofile}
\alias{config_Rprofile}
\title{Create or modify the \file{.Rprofile} file for a website project}
\usage{
config_Rprofile()
}
\value{
As a side-effect, the file \file{.Rprofile} is created or modified.
}
\description{
If the file \file{.Rprofile} does not exist in the current directory, copy
the file from the \file{resources} directory of \pkg{blogdown}. If the option
\code{blogdown.hugo.version} is not found in this file, append
\code{options(blogdown.hugo.version = "VERSION")} to it, where \code{VERSION}
is obtained from \code{\link{hugo_version}()}.
}
