% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{config_vercel}
\alias{config_vercel}
\title{Create the configuration file for Vercel}
\usage{
config_vercel(output = "vercel.json")
}
\arguments{
\item{output}{Path to the output file, or \code{NULL} to print the config.}
}
\description{
Create \file{vercel.json} that contains the Hugo version currently used.
}
\references{
Vercel: \url{https://vercel.com}
}
