% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dep_path}
\alias{dep_path}
\title{A helper function to return a dependency path name}
\usage{
dep_path(default = knitr::opts_chunk$get("fig.path"))
}
\arguments{
\item{default}{Return this default value when this function is called outside
of a \pkg{knitr} code chunk.}
}
\value{
A character string of the \code{default} value (outside \pkg{knitr}),
  or a path consisting of the \pkg{knitr} figure path appended by the current
  chunk label.
}
\description{
In most cases, \pkg{blogdown} can process images and HTML widgets
automatically generated from code chunks (they will be moved to the
\code{static/} folder by default), but it may fail to recognize dependency
files generated to other paths. This function returns a path that you can use
for your output files, so that \pkg{blogdown} knows that they should be be
processed, too. It is designed to be used in a \pkg{knitr} code chunk.
}
