% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{digest_if_exists}
\alias{digest_if_exists}
\title{Calculate the file digest if the file exists}
\usage{
digest_if_exists(file, alg = NA)
}
\arguments{
\item{file}{A string containing the path to a file.}

\item{alg}{A string containing the name of the digest algorithm to use.
Set the algorithm with `options(blogdown.digest = <algorithm>)`.
If no option was set, use crc32.}
}
\value{
A string: if the file exists return the digest. Otherwise, return
\code{NA}.
}
\description{
\code{digest_if_file_exists} returns a character vector with the hashed digest
of the file, or \code{NA} if the file does not exist.
}
\seealso{
\code{\link[digest:digest]{digest()}} for details about available
algorithms.
}
\keyword{internal}
