% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{digests}
\alias{digests}
\title{Digests for detecting file modifications.}
\description{
When creating digests for detecting file modifications, there is
a choice of algorithms
}
\details{
File digests are created by applying a hashing algorithm to the text contents
of a source file (\code{.Rmd} or  \code{.rmarkdown}).
By default, \code{blogdown} uses crc32 because it is fast and is good enough
for detecting changes in source files. This would not be good enough for
uniquely identifying a given file revision or source file (e.g., for revision
control), but that's not necessary for the purposes of updating sites in
`blogdown`.

The user can override the hashing algorithm by setting
\code{options(blogdown.algorithm = <algorithm>)},
where \code{<algorithm>} is one of the allowed digest algorithms
in \code{\link[digest]{digest}()}: "\code{crc32}", "\code{md5}",
"\code{sha1}", "\code{sha256}", "\code{sha512}",
"\code{xxhash32}", "\code{xxhash64}", or "\code{murmur32}".
}
\examples{
options(blogdown.algorithm = "sha256")
}
\seealso{
\code{\link{update_site_digests}()}, \code{\link{update_site}()}.
}
