% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{find_blog_content}
\alias{find_blog_content}
\title{Get the "content" directory of the site}
\usage{
find_blog_content()
}
\value{
A string containing the absolute path of the root directory for the
site.
}
\description{
`find_blog_content` gets the "content" directory of the site.
}
\keyword{internal}
