% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{find_hugo}
\alias{find_hugo}
\title{Find the Hugo executable}
\usage{
find_hugo(version = getOption("blogdown.hugo.version"))
}
\arguments{
\item{version}{The expected version number, e.g., \code{'0.25.1'}. If
\code{NULL}, it will try to find the maximum possible version.}
}
\value{
The path to the Hugo executable if found, otherwise it will signal an
  error, with a hint on how to install (the required version of) Hugo.
}
\description{
Search for Hugo in a series of possible installation directories (see
\code{\link{install_hugo}()} for these directories).
}
