% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{get_digest_algorithm}
\alias{get_digest_algorithm}
\title{Get the digest algorithm to use}
\usage{
get_digest_algorithm()
}
\value{
A string containing the name of the algorithm.
}
\description{
`get_digest_algorithm` gets the digest algorithm that will be used.
}
\details{
Set the algorithm with `options(blogdown.algorithm = <algorithm>)`.
If the option is not set, then use crc32.
}
\keyword{internal}
