% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{md5sum_filter}
\alias{md5sum_filter}
\title{Filter files by checking if their MD5 checksums have changed}
\usage{
md5sum_filter(files, db = "blogdown/md5sum.txt")
}
\arguments{
\item{files}{A vector of file paths.}

\item{db}{Path to the database file.}
}
\value{
Paths of files of which the checksums have changed.
}
\description{
Read the MD5 checksums of files from a database (a tab-separated text file),
and return the files of which the checksums have changed. If the database
does not exist, write the checksums of files to it, otherwise update the
checksums after the changed files have been identified.
}
\details{
When a file is modified, its MD5 checksum is very likely to change. This
function can be used to determine which Rmd files to be rebuilt in a
\pkg{blogdown} website. See \code{\link{build_site}()} for more information.
}
