% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{prune_site_digests}
\alias{prune_site_digests}
\title{Delete stored digests for specified source files}
\usage{
prune_site_digests(files)
}
\arguments{
\item{files}{A character vector of paths to the source files to be removed.}
}
\value{
The path to the digest file.
}
\description{
\code{prune_site_digests} removes the lines from the digest file corresponding to
a vector of source files.
}
\details{
Modifies the stored digest file to remove lines corresponding to selected
source files.
}
\seealso{
\code{\link{update_site_digests}()}.
}
