% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_rmd_digests}
\alias{update_rmd_digests}
\title{Generates and stores digests for all source and output files.}
\usage{
update_rmd_digests(files, partial = FALSE)
}
\arguments{
\item{files}{A character vector of paths to the source files.}

\item{partial}{Logical. If `TRUE`, keep rows from digest file for source
files that aren't in `files`. Otherwise, get rid of the old file and only
keep digests for source files in `files`.}
}
\value{
The path to the digest file.
}
\description{
`update_rmd_digests` calculates hashed digests for a list of source files
and their corresponding output files and stores them in a file.
}
\details{
Generates new hashed digests for both source and destination (output) files
and save the digests to a file "`digests.Rds`" in the root directory of the
site.
}
\keyword{internal}
