% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_site}
\alias{update_site}
\title{Update all files that are out of date}
\usage{
update_site(dir = NULL, quiet = FALSE)
}
\arguments{
\item{dir}{A string containing the root directory for checking.
By default, the "content" directory of the project.}

\item{quiet}{Suppress output. By default this is \code{FALSE} and the
function emits an informational message about how many files will
be rebuit.}
}
\value{
This function does not return anything
}
\description{
\code{update_site} rebuilds all source files that are new or have changed since
the last time the site was built.
}
\details{
Given a source directory (by default the "content" directory in the
root directory of the project), find all source files (\code{.Rmd} and
\code{.rmarkdown}) in the directory tree under the source directory,
calculate hashed digests of the files, and compare them to a
stored list of digests from the last time the site was built.

If the digests of either the source or output files don't match,
if a source file is new since the last time the site was built,
or if the output file does not exist,
then render the source file.

After rendering any out-of-date files, regenerate the digest list
and saves it to a file.
}
\seealso{
\code{\link{build_site}()}, \code{\link{build_dir}()},
\code{\link{digests}}.
}
