% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_site_digests}
\alias{update_site_digests}
\title{Generates and stores digests for all source and output files.}
\usage{
update_site_digests(dir = NA, partial = FALSE)
}
\arguments{
\item{dir}{A string with the name of the directory to search
(by default the "content" directory at the top-level directory of the site)}

\item{partial}{Logical. If \code{TRUE}, keep digests for source
files that aren't in the specified directory and its children and
descendants.
Otherwise, get rid of the old digest file and only keep digests for
source files in the source directory and its descendants.}
}
\value{
The path to the digest file.
}
\description{
\code{update_site_digests} calculates hashed digests for a site.
}
\details{
Generates new hashed digests for both source and destination (output) files
and save the digests to a file "\code{digests.Rds}" in the root directory of the
site.
}
\seealso{
\code{\link{prune_site_digests}()}, \code{\link{update_site}()},
\code{\link{digests}}.
}
