% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{find_hugo}
\alias{find_hugo}
\title{Find the Hugo executable}
\usage{
find_hugo(version = getOption("blogdown.hugo.version"))
}
\arguments{
\item{version}{The expected version number, e.g., \code{'0.25.1'}. If
\code{NULL}, it will try to find the maximum possible version. If
\code{'all'}, find all possible versions.}
}
\value{
The path to the Hugo executable if found, otherwise it will signal an
  error, with a hint on how to install (the required version of) Hugo. If
  \code{version = 'all'}, return the paths of all versions of Hugo installed.
}
\description{
Search for Hugo in a series of possible installation directories (see
\code{\link{install_hugo}()} for these directories).
}
\details{
If your website depends on a specific version of Hugo, we strongly recommend
that you set \code{options(blogdown.hugo.version = )} to the version number
you desire in the file \code{.Rprofile} in the root directory of the website
project, so that \pkg{blogdown} can try to find the right version of Hugo
before it builds or serves the website. You can use the function
\code{\link{config_Rprofile}()} to do this automatically.
}
