% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{serve_site}
\alias{serve_site}
\title{Live preview a site}
\usage{
serve_site(...)
}
\arguments{
\item{...}{Arguments passed to \code{servr::httw()} (arguments \code{dir},
\code{site.dir}, \code{baseurl}, and \code{handler} have been provided,
hence you cannot customize these arguments).}
}
\description{
Calls \code{servr::\link{httw}()} to watch for changes in the site, rebuild
the site if necessary, and refresh the web page automatically.
}
